# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FANTA11 Fantasy Soccer API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  teamPoints: number;
  gameweekPoints: number;
  globalRank: number;
  leagueCount: number;
  playerCount: number;
  budgetRemaining: number;
  /** @nullable */
  captainName?: string | null;
  /** @nullable */
  captainPoints?: number | null;
  /** @nullable */
  topScorerName?: string | null;
  /** @nullable */
  topScorerPoints?: number | null;
}