# lib/api-zod/src/generated/types/joinLeagueInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FANTA11 Fantasy Soccer API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinLeagueInput {
  teamId: number;
  code?: string;
}