# lib/api-zod/src/generated/types/leagueInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FANTA11 Fantasy Soccer API
 * OpenAPI spec version: 0.1.0
 */

export interface LeagueInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}