# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FANTA11 Fantasy Soccer API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerPosition } from './playerPosition';

export interface Player {
  id: number;
  name: string;
  position: PlayerPosition;
  club: string;
  clubShortName?: string;
  totalPoints: number;
  price: number;
  form?: number;
  /** Percentage of teams who have this player */
  selected?: number;
  goalsScored?: number;
  assists?: number;
  cleanSheets?: number;
  /** @nullable */
  imageUrl?: string | null;
}