# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FANTA11 Fantasy Soccer API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  name: string;
  managerName: string;
  totalPoints: number;
  budget: number;
  /** @nullable */
  captainId?: number | null;
  /** @nullable */
  viceCaptainId?: number | null;
  createdAt?: string;
}